implement isAdmin() check for admin commands so we can move all commands to use same args
