+++ commands/admin/mount.js
@@ -9,6 +9,8 @@ module.exports = class Mount extends Command {
   }
 
   call(bot, opts) {
+    if (!this.adminCallable) return
+
     const module = opts.args[0]
     const cmd = CommandHandler.commandList[module]
     const obs = ObserverHandler.observerList[module]
