+++ commands/admins.js
@@ -1,13 +1,6 @@
 const Command = require('./command.js')
-const config = require("../config.js");
-const _ = require('lodash');
+const config = require("../config.js")
 
-/*
- * Determine if a supplied user (by nickname) is an admin for the bot as defined
- * in the config file.
- *
- * !admin user123
- */
 module.exports = class Admins extends Command {
 
   constructor() {
@@ -24,16 +17,6 @@ module.exports = class Admins extends Command {
     }
     
     return respond(`${admins}. Global admins: ${globalAdmins}`)    
-  }
-
-  // Check if a user is an admin for a specified channel
-  static isAdmin(user, channel) {
-    if (_.includes(config.globalAdmins, user)) {
-      return true
-    } else if (config.admin[channel]) {
-      return _.includes(config.admin[channel], user)
-    }
-    return false
   }
 
 }
